+++ src/components/NewExpense/ExpenseCategory.css
@@ -1,0 +1,5 @@
+.category-container {
+  display: flex;
+  flex-direction: column;
+  row-gap: 0.5rem;
+}
