+++ src/components/NewExpense/ExpenseDescription.css
@@ -1,0 +1,14 @@
+.description-container > label {
+  display: inline-block;
+  margin-bottom: 0.5rem;
+}
+textarea {
+  border: none;
+  background: #f5f5f5;
+  height: 150px;
+  font-size: 1rem;
+  padding: 1rem;
+  resize: none;
+  width: 100%;
+  box-sizing: border-box;
+}
