+++ src/components/NewExpense/ExpenseDescription.js
@@ -1,0 +1,23 @@
+import { useState } from "react";
+import "./ExpenseDescription.css";
+function ExpenseDescription() {
+  const [description, setDescription] = useState("");
+  function descriptionChangeHandler(event) {
+    setDescription(event.target.value.trim());
+  }
+
+  return (
+    <div className="description-container">
+      <label htmlFor="description">Description</label>
+      <textarea
+        name="description"
+        placeholder="Description (optional)"
+        id="description"
+        onChange={descriptionChangeHandler}
+        value={description}
+      ></textarea>
+    </div>
+  );
+}
+
+export default ExpenseDescription;
